Move StringUtil to util package
