+++ backend/src/main/java/com/empire/util/StringUtil.java
@@ -1,4 +1,4 @@
-package com.empire;
+package com.empire.util;
 
 import java.util.List;
